# EventHub-BE/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "EventHub BE",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5678",
    "basePath": "/",
    "paths": {
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get metrics and overview data for the admin dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get admin dashboard information"
                ],
                "summary": "Get admin dashboard data",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AdminDashboardResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/communities/popular": {
            "get": {
                "description": "Get list of popular communities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "popular communities"
                ],
                "summary": "Get popular communities",
                "responses": {
                    "200": {
                        "description": "popular communities retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.PopularCommunity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
        "/communities/{id}": {
            "get": {
                "description": "Get detailed information of a specific community by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities detail"
                ],
                "summary": "Get community detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "community detail retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.CommunityDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid community id",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/events/changeuserprofile": {
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Update user profile details for the authenticated user",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Change User Profile"
                ],
                "summary": "Change user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User bio",
                        "name": "bio",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "User location",
                        "name": "location",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "User job",
                        "name": "job",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Profile image",
                        "name": "profile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "profile updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "bad request error message",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "401": {
                        "description": "invalid token",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
        "/events/createEvent": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Create a new event using JSON payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "create events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "Create event payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "event created success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "failed to bind",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
        "/events/{id}": {
            "get": {
                "description": "Get detailed information of a specific event by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "event detail retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EventDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid event id",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdminDashboardResponse": {
            "type": "object",
            "properties": {
                "avg_fill_rate": {
                    "type": "number"
                },
                "total_communities": {
                    "type": "integer"
                },
                "total_events": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateEvent": {
            "type": "object",
            "required": [
                "capacity",
                "category_id",
                "description",
                "end_time",
                "event_format",
                "images",
                "location",
                "start_time",
                "title"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "category_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "community_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "event_format": {
                    "type": "string"
                },
                "images": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "position_job": {
                    "type": "string"
                },
                "speaker_id": {
                    "type": "integer"
                },
                "speaker_name": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.Res": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "token": {}
            }
        },
        "model.CommunityDetail": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "category_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "images": {
                    "type": "string"
                },
                "member_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.EventDetail": {
            "type": "object",
            "properties": {
                "attendees": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "community_id": {
                    "type": "integer"
                },
                "community_images": {
                    "type": "string"
                },
                "community_title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "event_format": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.PopularCommunity": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "string"
                },
                "member_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}